fix: resolve 50+ bugs across codebase — critical crashes, logic errors, f-string gaps
CRITICAL:
- lock_monitor: fix circular class inheritance, unlock event pass-statement, get_session_id
- welcome_briefing: wrong ScreenMonitor class → ScreenIntelligence
- screen_monitor: take_screenshot path param ignored (mss.shot args)
- branding: float coords to PIL rectangle (wrap in int)
- gesture_control: wrong MediaPipe field names (min_hand_detection → min_detection)
- workflow: threading.Timer args=[list] → args=(tuple,) positional arg fix
- system_adapter: missing tempfile import (NameError on URL install)
- cmd_control: select module Unix-only → msvcrt for Windows
- config_manager: infinite recursion in is_configured()
- rag_pipeline: wrong Gemini API call (generate → models.generate_content)

HIGH (85+ f-string fixes):
- 85+ logging calls missing f-prefix across actions/, core/, integrations/, memory/
  Files: game_updater, screen_processor, youtube_video, computer_settings,
  web_search, send_message, computer_control, code_helper, claude_shell,
  flight_finder, github_tool, dev_agent, memory_manager, error_handler,
  conversation_manager, scheduler, plugin_manager, startup_launcher, fix_shortcut
- outlook_adapter: implement stub _action_update/delete/find_meeting_time
- llm_orchestrator: resolve_value loop to fix all ${steps[N].result.field} substitutions
- universal_orchestrator: remove overly broad alias substring containment check
- memory_manager: save_memory() now writes passed arg not stale _pending_writes

MEDIUM:
- gesture_control: camera release on early return
- screen_watchdog: _last_window_title update on keyword alert
- intro_music: _autopan() now applies pan value (was computing and discarding)
- computer_control: _clipboard_copy returns empty message vs empty string
- whatsapp_adapter: class-level dict mutation → __init__ (import-time side effect)
- llm_orchestrator: _resolve_value() resolves ALL substitutions (while loop)

LOW:
- conversation_context: stub functions add debug logging

+++ core/branding.py
@@ -47,8 +47,8 @@ def _draw_j_monogram(size: int, style: str = "default") -> Image.Image:
         stem_x = cx + size * 0.08
         # Vertical stem
         draw.rectangle(
-            [stem_x - stem_w/2, cy - size * 0.25,
-             stem_x + stem_w/2, cy + size * 0.25],
+            [int(stem_x - stem_w/2), int(cy - size * 0.25),
+             int(stem_x + stem_w/2), int(cy + size * 0.25)],
             fill=BRAND_CYAN
         )
         # Bottom curve (thick arc)
